HAI-1571 Add possibility to auto generate hanke on hakemus (#283)
User must be able to create an application (hakemus) without existing hanke. However, an application must always have a related hanke. This commit adds a feature to auto-generate a hanke for johtoselvitys application.

When the user decides to update any information of the autogenerated hanke, it must contain the required information / fulfil same requirements as with a normal hanke.

+++ services/hanke-service/src/main/kotlin/fi/hel/haitaton/hanke/application/ApplicationController.kt
@@ -6,6 +6,8 @@ import fi.hel.haitaton.hanke.HankeService
 import fi.hel.haitaton.hanke.currentUserId
 import fi.hel.haitaton.hanke.logging.DisclosureLogService
 import fi.hel.haitaton.hanke.permissions.PermissionCode
+import fi.hel.haitaton.hanke.permissions.PermissionCode.EDIT_APPLICATIONS
+import fi.hel.haitaton.hanke.permissions.PermissionCode.VIEW
 import fi.hel.haitaton.hanke.permissions.PermissionService
 import fi.hel.haitaton.hanke.validation.InvalidApplicationDataException
 import fi.hel.haitaton.hanke.validation.ValidApplication
@@ -18,7 +20,9 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses
 import mu.KotlinLogging
 import org.springframework.http.HttpHeaders
 import org.springframework.http.HttpStatus
-import org.springframework.http.MediaType
+import org.springframework.http.MediaType.APPLICATION_JSON_VALUE
+import org.springframework.http.MediaType.APPLICATION_PDF
+import org.springframework.http.MediaType.APPLICATION_PDF_VALUE
 import org.springframework.http.ResponseEntity
 import org.springframework.validation.annotation.Validated
 import org.springframework.web.bind.annotation.DeleteMapping
@@ -73,7 +77,7 @@ class ApplicationController(
             ]
     )
     fun getById(@PathVariable(name = "id") id: Long): Application {
-        checkHakemusOikeus(id, PermissionCode.VIEW)
+        checkHakemusPermission(id, VIEW)
         val userId = currentUserId()
         val application = service.getApplicationById(id)
         disclosureLogService.saveDisclosureLogsForApplication(application, userId)
@@ -101,19 +105,40 @@ class ApplicationController(
     fun create(@ValidApplication @RequestBody application: Application): Application {
         val userId = currentUserId()
         val hankeTunnus = application.hankeTunnus
-        val hankeId = hankeService.getHankeId(hankeTunnus)
 
-        if (
-            hankeId == null ||
-                !permissionService.hasPermission(hankeId, userId, PermissionCode.EDIT_APPLICATIONS)
-        ) {
-            throw HankeNotFoundException(hankeTunnus)
-        }
+        checkPermissionToCreate(hankeTunnus, userId)
 
         val createdApplication = service.create(application, userId)
 
         disclosureLogService.saveDisclosureLogsForApplication(createdApplication, userId)
         return createdApplication
+    }
+
+    @PostMapping("/johtoselvitys")
+    @Operation(
+        summary = "Generates new hanke from cable report application and saves application to it.",
+        description =
+            "Returns the created application. The new application is created as a draft, " +
+                "i.e. with true in pendingOnClient. The draft is not sent to Allu. "
+    )
+    @ApiResponses(
+        value =
+            [
+                ApiResponse(description = "The created application", responseCode = "200"),
+                ApiResponse(
+                    description = "The request body was invalid",
+                    responseCode = "400",
+                    content = [Content(schema = Schema(implementation = HankeError::class))]
+                ),
+            ]
+    )
+    fun createWithGeneratedHanke(@RequestBody cableReport: CableReportWithoutHanke): Application {
+        val userId = currentUserId()
+        return hankeService
+            .generateHankeWithApplication(cableReport, userId)
+            .applications
+            .first()
+            .also { disclosureLogService.saveDisclosureLogsForApplication(it, userId) }
     }
 
     @PutMapping("/{id}")
@@ -154,7 +179,7 @@ class ApplicationController(
         @PathVariable(name = "id") id: Long,
         @ValidApplication @RequestBody application: Application
     ): Application {
-        checkHakemusOikeus(id, PermissionCode.EDIT_APPLICATIONS)
+        checkHakemusPermission(id, EDIT_APPLICATIONS)
         val userId = currentUserId()
         val updatedApplication =
             service.updateApplicationData(id, application.applicationData, userId)
@@ -190,7 +215,7 @@ class ApplicationController(
             ]
     )
     fun delete(@PathVariable(name = "id") id: Long) {
-        checkHakemusOikeus(id, PermissionCode.EDIT_APPLICATIONS)
+        checkHakemusPermission(id, EDIT_APPLICATIONS)
         service.delete(id, currentUserId())
     }
 
@@ -223,14 +248,11 @@ class ApplicationController(
             ]
     )
     fun sendApplication(@PathVariable(name = "id") id: Long): Application {
-        checkHakemusOikeus(id, PermissionCode.EDIT_APPLICATIONS)
+        checkHakemusPermission(id, EDIT_APPLICATIONS)
         return service.sendApplication(id, currentUserId())
     }
 
-    @GetMapping(
-        "/{id}/paatos",
-        produces = [MediaType.APPLICATION_PDF_VALUE, MediaType.APPLICATION_JSON_VALUE]
-    )
+    @GetMapping("/{id}/paatos", produces = [APPLICATION_PDF_VALUE, APPLICATION_JSON_VALUE])
     @Operation(
         summary = "Download a decision",
         description = "Downloads a decision for this application from Allu."
@@ -248,18 +270,15 @@ class ApplicationController(
             ]
     )
     fun downloadDecision(@PathVariable(name = "id") id: Long): ResponseEntity<ByteArray> {
-        checkHakemusOikeus(id, PermissionCode.VIEW)
+        checkHakemusPermission(id, VIEW)
         val (filename, pdfBytes) = service.downloadDecision(id, currentUserId())
 
         val headers = HttpHeaders()
         headers.add("Content-Disposition", "inline; filename=$filename.pdf")
-        return ResponseEntity.ok()
-            .headers(headers)
-            .contentType(MediaType.APPLICATION_PDF)
-            .body(pdfBytes)
+        return ResponseEntity.ok().headers(headers).contentType(APPLICATION_PDF).body(pdfBytes)
     }
 
-    fun checkHakemusOikeus(hakemusId: Long, permissionCode: PermissionCode) {
+    fun checkHakemusPermission(hakemusId: Long, permissionCode: PermissionCode) {
         val userId = currentUserId()
         val application = service.getApplicationById(hakemusId)
         val hankeId = hankeService.getHankeId(application.hankeTunnus)
@@ -267,6 +286,12 @@ class ApplicationController(
         if (hankeId == null || !permissionService.hasPermission(hankeId, userId, permissionCode)) {
             throw ApplicationNotFoundException(hakemusId)
         }
+    }
+
+    fun checkPermissionToCreate(hankeTunnus: String, userId: String) {
+        val id = hankeService.getHankeId(hankeTunnus)
+        if (id == null || !permissionService.hasPermission(id, userId, EDIT_APPLICATIONS))
+            throw HankeNotFoundException(hankeTunnus)
     }
 
     @ExceptionHandler(ApplicationNotFoundException::class)
